AzureMonitor: Remove deprecated configuration parameters
