+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/ConfigEditor.tsx
@@ -1,21 +1,15 @@
-import React, { PureComponent } from 'react';
-import {
-  DataSourcePluginOptionsEditorProps,
-  SelectableValue,
-  updateDatasourcePluginJsonDataOption,
-  updateDatasourcePluginOption,
-  updateDatasourcePluginResetOption,
-  updateDatasourcePluginSecureJsonDataOption,
-} from '@grafana/data';
-import { Alert } from '@grafana/ui';
-import { MonitorConfig } from './MonitorConfig';
-import { AnalyticsConfig } from './AnalyticsConfig';
+import { DataSourcePluginOptionsEditorProps, SelectableValue, updateDatasourcePluginOption } from '@grafana/data';
 import { getBackendSrv, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
-import { InsightsConfig } from './InsightsConfig';
+import { Alert } from '@grafana/ui';
+import React, { PureComponent } from 'react';
+
 import ResponseParser from '../azure_monitor/response_parser';
 import { AzureDataSourceJsonData, AzureDataSourceSecureJsonData, AzureDataSourceSettings } from '../types';
-import { isAppInsightsConfigured } from '../credentials';
 import { routeNames } from '../utils/common';
+import { AnalyticsConfig } from './deprecated/components/AnalyticsConfig';
+import { InsightsConfig } from './deprecated/components/InsightsConfig';
+import { gtGrafana9, isAppInsightsConfigured } from './deprecated/utils';
+import { MonitorConfig } from './MonitorConfig';
 
 export type Props = DataSourcePluginOptionsEditorProps<AzureDataSourceJsonData, AzureDataSourceSecureJsonData>;
 
@@ -27,7 +21,6 @@ interface ErrorMessage {
 
 export interface State {
   unsaved: boolean;
-  appInsightsInitiallyConfigured: boolean;
   error?: ErrorMessage;
 }
 
@@ -40,7 +33,6 @@ export class ConfigEditor extends PureComponent<Props, State> {
 
     this.state = {
       unsaved: false,
-      appInsightsInitiallyConfigured: isAppInsightsConfigured(props.options),
     };
     this.baseURL = `/api/datasources/${this.props.options.id}/resources/${routeNames.azureMonitor}/subscriptions`;
   }
@@ -90,24 +82,6 @@ export class ConfigEditor extends PureComponent<Props, State> {
     }
   };
 
-  // TODO: Used only by InsightsConfig
-  private onUpdateJsonDataOption =
-    (key: keyof AzureDataSourceJsonData) => (event: React.SyntheticEvent<HTMLInputElement | HTMLSelectElement>) => {
-      updateDatasourcePluginJsonDataOption(this.props, key, event.currentTarget.value);
-    };
-
-  // TODO: Used only by InsightsConfig
-  private onUpdateSecureJsonDataOption =
-    (key: keyof AzureDataSourceSecureJsonData) =>
-    (event: React.SyntheticEvent<HTMLInputElement | HTMLSelectElement>) => {
-      updateDatasourcePluginSecureJsonDataOption(this.props, key, event.currentTarget.value);
-    };
-
-  // TODO: Used only by InsightsConfig
-  private resetSecureKey = (key: keyof AzureDataSourceSecureJsonData) => {
-    updateDatasourcePluginResetOption(this.props, key);
-  };
-
   render() {
     const { options } = this.props;
     const { error } = this.state;
@@ -115,16 +89,14 @@ export class ConfigEditor extends PureComponent<Props, State> {
     return (
       <>
         <MonitorConfig options={options} updateOptions={this.updateOptions} getSubscriptions={this.getSubscriptions} />
-        <AnalyticsConfig options={options} updateOptions={this.updateOptions} />
-        {this.state.appInsightsInitiallyConfigured && (
-          <InsightsConfig
-            options={options}
-            onUpdateJsonDataOption={this.onUpdateJsonDataOption}
-            onUpdateSecureJsonDataOption={this.onUpdateSecureJsonDataOption}
-            onResetOptionKey={this.resetSecureKey}
-          />
+        {/* Remove with Grafana 9 */}
+        {!gtGrafana9() && (
+          <>
+            <AnalyticsConfig options={options} updateOptions={this.updateOptions} />
+            {isAppInsightsConfigured(options) && <InsightsConfig {...this.props} />}
+          </>
         )}
-
+        {/* ===================== */}
         {error && (
           <Alert severity="error" title={error.title}>
             <p>{error.description}</p>
