+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/credentials.ts
@@ -1,4 +1,5 @@
 import { config } from '@grafana/runtime';
+
 import {
   AzureAuthType,
   AzureCloud,
@@ -164,8 +165,4 @@ export function updateCredentials(
 
       return options;
   }
-}
-
-export function isAppInsightsConfigured(options: AzureDataSourceSettings) {
-  return !!(options.jsonData.appInsightsAppId && options.secureJsonFields.appInsightsApiKey);
 }
